refactor: extract KSC action reconciler
